research-service/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.9.2
faiss-cpu==1.9.0
numpy==2.1.3
pypdf==5.1.0
python-multipart==0.0.17
requests==2.32.3
